# bridge/apps/backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/clients/heartbeat": {
      "post": {
        "operationId": "ClientsController_heartbeat",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClientHeartbeatDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        }
      }
    },
    "/api/v1/clients/events": {
      "post": {
        "operationId": "ClientsController_event",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClientEventDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        }
      }
    },
...
  },
  "info": {
    "title": "ntlstl",
    "description": "ntlstl API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ClientHeartbeatDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 255
          },
          "kind": {
            "type": "string",
            "enum": [
              "reports"
            ]
          }
        },
        "required": [
          "name",
          "kind"
        ]
      },
      "ClientEventDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 255
          },
          "type": {
            "type": "string",
            "enum": [
              "pulled",
              "pull_failed",
              "issues_created"
            ]
          },
          "taskKey": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "branch": {
            "type": "string",
            "maxLength": 200
          },
          "error": {
            "type": "string",
            "maxLength": 300
          },
          "count": {
            "type": "number",
            "minimum": 0,
            "maximum": 1000
          }
        },
        "required": [
          "name",
          "type",
          "taskKey"
        ]
      },
...
    }
  }
}